requirements/base.txt
amqp==5.1.1
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==22.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.32.0
cloudscraper==1.2.69
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
cryptography==39.0.2
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-celery-beat==2.5.0
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==23.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
drf-yasg==1.21.4
exceptiongroup==1.0.4
Flask==2.2.3
Flask-Cors==3.0.10
flask-restplus==0.13.0
flower==1.2.0
humanize==4.6.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==22.0
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
pytest-django==4.5.2
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0
redis==4.5.4
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
stripe==5.4.0
swagger-ui==0.1.2
tomli==2.0.1
tornado==6.3
typing==3.7.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
whitenoise==6.2.0
